# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hostel360 API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/finance/rent-payments": {
      "put": {
        "tags": [
          "finance-controller"
        ],
        "operationId": "setRentPaid",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/finance/year/{year}": {
      "get": {
        "tags": [
          "finance-controller"
        ],
        "operationId": "financeYear",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RentPaymentRequest": {
        "type": "object",
        "properties": {
          "stayId": {
            "type": "integer",
            "format": "int64"
          },
          "month": {
            "type": "string",
            "format": "date"
          },
          "paid": {
            "type": "boolean"
          }
        },
        "required": [
          "month",
          "paid",
          "stayId"
        ]
      },
...
      "CategoryTotal": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "CLEANING",
              "LAUNDRY",
              "REPAIRS",
              "ELECTRICITY",
              "WATER",
              "HEATING",
              "INTERNET",
              "GARBAGE",
              "STAFF",
              "TAXES",
              "EQUIPMENT",
              "OTHER"
            ]
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "amount",
          "category"
        ]
      },
      "MonthSummary": {
        "type": "object",
        "description": "Income counts on the arrival date; long-term rent counts once per month.",
        "properties": {
          "month": {
            "type": "string",
            "format": "date",
            "description": "First day of the month"
          },
          "booking": {
            "type": "number"
          },
          "direct": {
            "type": "number"
          },
          "longTerm": {
            "type": "number"
          },
          "income": {
            "type": "number"
          },
          "commission": {
            "type": "number"
          },
          "expenses": {
            "type": "number"
          },
          "costs": {
            "type": "number"
          },
          "profit": {
            "type": "number"
          },
          "byCategory": {
            "type": "array",
            "description": "Expenses by category, largest first",
            "items": {
              "$ref": "#/components/schemas/CategoryTotal"
            }
          },
          "arrivals": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "arrivals",
          "booking",
          "byCategory",
          "commission",
          "costs",
          "direct",
          "expenses",
          "income",
          "longTerm",
          "month",
          "profit"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}